+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added Laravel 13.x support
 - Added facades for parsers (`WKBParser`, `WKTParser`, `GeojsonParser`) and generators (`WKBGenerator`, `WKTGenerator`, `GeojsonGenerator`)
 
 ### Fixed